move mermaid pkg to internal
