+++ clab/graph.go
@@ -15,8 +15,8 @@ import (
 	"github.com/awalterschulze/gographviz"
 	log "github.com/sirupsen/logrus"
 	e "github.com/srl-labs/containerlab/errors"
+	"github.com/srl-labs/containerlab/internal/mermaid"
 	"github.com/srl-labs/containerlab/labels"
-	"github.com/srl-labs/containerlab/mermaid"
 	"github.com/srl-labs/containerlab/nodes"
 	"github.com/srl-labs/containerlab/runtime"
 	"github.com/srl-labs/containerlab/types"
